pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "idk"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentevals>=0.0.9",
    "aiohttp>=3.13.2",
    "asyncio>=4.0.0",
    "boto3>=1.42.19",
    "deepeval>=3.7.8",
    "deepevals>=0.2.0",
    "fastembed>=0.7.3",
    "google-genai>=1.56.0",
    "jupyter>=1.1.1",
    "langchain>=1.0.7",
    "langchain-core>=1.2.4,<1.2.6",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3,<4.1.3",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "langchain-voyageai>=0.3.0",
    "langgraph>=1.0.3,<1.0.5",
    "langgraph-cli[inmem]>=0.4.7",
    "litellm>=1.80.11",
    "onnxruntime>=1.23.2",
    "openai>=2.14.0",
    "opencv-python-headless>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=11.0.0,<12.0.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.11.3",
    "ragas>=0.4.1",
    "requests>=2.32.5",
    "scikit-network>=0.33.5",
    "scipy>=1.16.3",
    "supabase>=2.27.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "opik>=1.9.73",
]

[project.optional-dependencies]
torch-cpu = [
    "torch>=2.0.0; platform_machine != 'arm64'",
]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

